Sample parameters from approximate Gaussian posterior distribution

Efficient Monte Carlo sampling of parameters (and REPORTed quantities) from the approximate posterior of an RTMB model. See sdreport for details on posterior variance-covariance in random effects models.

Usage

mcreport(
  obj,
  nSamples = 1000,
  include_random_pars = TRUE,
  report = FALSE,
  Q = NULL,
  ...
)

Arguments

obj

Optimised RTMB object generated by MakeADFun

nSamples

Number of samples to draw

include_random_pars

Logical; Should random parameters be included in the output?

report

Logical; Should REPORTed quantities be sampled as well? Defaults to FALSE because this may be slow depending on your model.

Q

Optional precalculated sparse precision matrix returned by sdreport(..., getJointPrecision = TRUE)$jointPrecision. If not provided, computed internally using sdreport. Only used for models with random effects.

...

For internal use only

Value

A list structured like the original parameter list used in the MakeADFun call (potentially including additional REPORTed quantities). Each entry is a list with nSamples entries.

Details

Caution: This has nothing to do with Bayesian posterior sampling. It is simple Monte Carlo sampling from a Gaussian distribution with mean at the MLE and covariance given by the inverse of the Hessian (for fixed effects models) or the joint precision matrix (for random effects models). This is a common approach to get an approximate idea of parameter uncertainty around the MLE, but it relies on large-sample asymptotics.

Sampling random effects from their posterior as compared to calculating marginal standard devitions like sdreport is particularly useful for multidimensional random effects (e.g. for locations \(x\) and \(y\)) where pointwise confidence intervals (e.g. along a path) based on standard deviations are not possible.

Examples

set.seed(123)

### Example 1: Without random effects ###

# simulate data
x <- rgumbel(100, location = 5, scale = 2)

# negative log-likelihood function
nll <- function(par) {
  getAll(par)
  x <- OBS(x) # mark x as the response
  scale <- exp(log_scale)
  ADREPORT(scale); REPORT(scale)
  -sum(dgumbel(x, loc, scale, log = TRUE))
}

# RTMB AD object
par <- list(loc = 5, log_scale = log(2))
obj <- MakeADFun(nll, par, silent = TRUE)

# model fitting using AD gradient
opt <- nlminb(obj$par, obj$fn, obj$gr)

# sample from asymptotic distribution of the MLE
samples <- mcreport(obj, report = TRUE)
#> Evaluating Hessian...
#> Computing reported quantities...
# paramaters
mean(unlist(samples$loc)); sd(unlist(samples$loc))
#> [1] 5.00983
#> [1] 0.1991763
# reported transformed parameters
mean(unlist(samples$scale)); sd(unlist(samples$scale))
#> [1] 1.970023
#> [1] 0.1544266

# compare to delta method sd from sdreport
sdr <- sdreport(obj)
summary(sdr)
#>            Estimate Std. Error
#> loc       5.0015427 0.20659355
#> log_scale 0.6732893 0.07663174
#> scale     1.9606760 0.15025002

### Example 2: With random effects ###

# simulate random effects
id <- rep(1:10, each = 100) # 10 groups with 100 observations each
true_gamma <- rnorm(10, 0, 2) # random coefficients
true_probs <- plogis(1 + true_gamma) # linear predictor

# simulate Bernoulli data conditional on random coefficients
x <- rbinom(1000, 1, true_probs[id])

# negative log-likelihood function
nll <- function(par) {
  getAll(par)
  x <- OBS(x) # mark x as the response

  # random effect likelihood
  sd_gamma <- exp(log_sd_gamma)
  ADREPORT(sd_gamma); REPORT(sd_gamma)
  nll <- -sum(dnorm(gamma, 0, sd_gamma, log = TRUE))

  # data likelihood
  probs <- plogis(beta0 + gamma) # linear predictor
  ADREPORT(probs); REPORT(probs)
  nll <- nll - sum(dbinom(x, 1, probs[id], log = TRUE))
  return(nll)
}

# RTMB Laplace approximation
par <- list(beta0 = 1, log_sd_gamma = log(2), gamma = rep(0,10))
obj <- MakeADFun(nll, par, random = "gamma")

# model fitting using AD gradient
opt <- nlminb(obj$par, obj$fn, obj$gr)
#> iter: 1  value: 466.6571 mgc: 25.89414 ustep: 1 
#> iter: 2  value: 458.8619 mgc: 7.783187 ustep: 1 
#> iter: 3  value: 457.9038 mgc: 2.314415 ustep: 1 
#> iter: 4  value: 457.8477 mgc: 0.496922 ustep: 1 
#> iter: 5  value: 457.8473 mgc: 0.04143599 ustep: 1 
#> iter: 6  value: 457.8473 mgc: 0.0003582528 ustep: 1 
#> iter: 7  value: 457.8473 mgc: 2.737444e-08 ustep: 1 
#> iter: 8  mgc: 1.315614e-14 
#> iter: 1  mgc: 1.315614e-14 
#> Matching hessian patterns... Done
#> outer mgc:  4.99482 
#> iter: 1  value: 455.6624 mgc: 7.257169 ustep: 1 
#> iter: 2  value: 454.8264 mgc: 2.830034 ustep: 1 
#> iter: 3  value: 454.8157 mgc: 0.3392154 ustep: 1 
#> iter: 4  value: 454.8157 mgc: 0.006531764 ustep: 1 
#> iter: 5  value: 454.8157 mgc: 2.531136e-06 ustep: 1 
#> iter: 6  mgc: 3.78364e-13 
#> iter: 1  mgc: 3.78364e-13 
#> outer mgc:  4.523853 
#> iter: 1  value: 464.6336 mgc: 19.65283 ustep: 1 
#> iter: 2  value: 456.8745 mgc: 7.303745 ustep: 1 
#> iter: 3  value: 456.8412 mgc: 0.5108803 ustep: 1 
#> iter: 4  value: 456.8412 mgc: 0.004387372 ustep: 1 
#> iter: 5  value: 456.8412 mgc: 4.812511e-07 ustep: 1 
#> iter: 6  mgc: 8.437695e-15 
#> iter: 1  value: 454.2411 mgc: 6.474134 ustep: 1 
#> iter: 2  value: 454.2219 mgc: 0.3670008 ustep: 1 
#> iter: 3  value: 454.2219 mgc: 0.001768257 ustep: 1 
#> iter: 4  value: 454.2219 mgc: 5.308701e-08 ustep: 1 
#> mgc: 1.154632e-14 
#> iter: 1  mgc: 1.154632e-14 
#> outer mgc:  0.6986839 
#> iter: 1  value: 454.7849 mgc: 6.485323 ustep: 1 
#> iter: 2  value: 454.7346 mgc: 0.4341287 ustep: 1 
#> iter: 3  value: 454.7345 mgc: 0.005420217 ustep: 1 
#> iter: 4  value: 454.7345 mgc: 2.08349e-06 ustep: 1 
#> iter: 5  mgc: 3.099743e-13 
#> iter: 1  value: 454.1972 mgc: 1.090947 ustep: 1 
#> iter: 2  value: 454.1972 mgc: 0.009769479 ustep: 1 
#> iter: 3  value: 454.1972 mgc: 1.859163e-06 ustep: 1 
#> iter: 4  mgc: 1.279532e-13 
#> iter: 1  mgc: 1.279532e-13 
#> outer mgc:  0.2913437 
#> iter: 1  value: 454.2338 mgc: 1.013011 ustep: 1 
#> iter: 2  value: 454.2337 mgc: 0.0079488 ustep: 1 
#> iter: 3  value: 454.2337 mgc: 1.019005e-06 ustep: 1 
#> iter: 4  mgc: 3.608225e-14 
#> iter: 1  mgc: 3.608225e-14 
#> outer mgc:  0.2254721 
#> iter: 1  value: 454.1926 mgc: 0.1157754 ustep: 1 
#> iter: 2  value: 454.1926 mgc: 0.00130599 ustep: 1 
#> iter: 3  value: 454.1926 mgc: 1.619449e-07 ustep: 1 
#> iter: 4  mgc: 9.65894e-15 
#> iter: 1  mgc: 9.65894e-15 
#> outer mgc:  0.1946059 
#> iter: 1  value: 454.2073 mgc: 0.1411521 ustep: 1 
#> iter: 2  value: 454.2073 mgc: 0.0004344698 ustep: 1 
#> iter: 3  value: 454.2073 mgc: 1.828603e-08 ustep: 1 
#> mgc: 1.132427e-14 
#> iter: 1  mgc: 1.132427e-14 
#> outer mgc:  0.02203839 
#> iter: 1  value: 454.2057 mgc: 0.0734571 ustep: 1 
#> iter: 2  value: 454.2057 mgc: 4.119246e-05 ustep: 1 
#> iter: 3  mgc: 2.812395e-11 
#> iter: 1  mgc: 2.812395e-11 
#> outer mgc:  0.003305415 
#> iter: 1  value: 454.2061 mgc: 0.01035708 ustep: 1 
#> iter: 2  value: 454.2061 mgc: 8.01635e-07 ustep: 1 
#> iter: 3  mgc: 1.64313e-14 
#> iter: 1  mgc: 1.64313e-14 
#> outer mgc:  1.362425e-05 
#> iter: 1  value: 454.2061 mgc: 1.796294e-05 ustep: 1 
#> iter: 2  mgc: 2.229716e-12 
#> iter: 1  mgc: 2.229716e-12 
#> outer mgc:  1.12991e-08 
#> iter: 1  mgc: 2.229716e-12 

# draw samples
samples <- mcreport(obj, report = TRUE)
#> Computing joint precision...
#> iter: 1  mgc: 2.229716e-12 
#> outer mgc:  1.12991e-08 
#> iter: 1  value: 454.2064 mgc: 0.02352665 ustep: 1 
#> iter: 2  value: 454.2064 mgc: 4.305452e-06 ustep: 1 
#> iter: 3  mgc: 3.10918e-13 
#> outer mgc:  0.00852029 
#> iter: 1  value: 454.2058 mgc: 0.02353236 ustep: 1 
#> iter: 2  value: 454.2058 mgc: 4.300467e-06 ustep: 1 
#> iter: 3  mgc: 3.338996e-13 
#> outer mgc:  0.008520353 
#> iter: 1  value: 454.2074 mgc: 0.003501856 ustep: 1 
#> iter: 2  value: 454.2074 mgc: 1.187864e-06 ustep: 1 
#> iter: 3  mgc: 1.338929e-13 
#> outer mgc:  0.01353103 
#> iter: 1  value: 454.2047 mgc: 0.003508867 ustep: 1 
#> iter: 2  value: 454.2047 mgc: 1.190836e-06 ustep: 1 
#> iter: 3  mgc: 1.39222e-13 
#> outer mgc:  0.01354524 
#> Computing reported quantities...

# run sdreport
sdr <- sdreport(obj)
#> iter: 1  mgc: 2.229716e-12 
#> outer mgc:  1.12991e-08 
#> iter: 1  value: 454.2064 mgc: 0.02352665 ustep: 1 
#> iter: 2  value: 454.2064 mgc: 4.305452e-06 ustep: 1 
#> iter: 3  mgc: 3.10918e-13 
#> outer mgc:  0.00852029 
#> iter: 1  value: 454.2058 mgc: 0.02353236 ustep: 1 
#> iter: 2  value: 454.2058 mgc: 4.300467e-06 ustep: 1 
#> iter: 3  mgc: 3.338996e-13 
#> outer mgc:  0.008520353 
#> iter: 1  value: 454.2074 mgc: 0.003501856 ustep: 1 
#> iter: 2  value: 454.2074 mgc: 1.187864e-06 ustep: 1 
#> iter: 3  mgc: 1.338929e-13 
#> outer mgc:  0.01353103 
#> iter: 1  value: 454.2047 mgc: 0.003508867 ustep: 1 
#> iter: 2  value: 454.2047 mgc: 1.190836e-06 ustep: 1 
#> iter: 3  mgc: 1.39222e-13 
#> outer mgc:  0.01354524 
#> outer mgc:  1.032585 

# standard deviation using delta method
as.list(sdr, "Std", report = TRUE)$probs
#>  [1] 0.04426637 0.04799042 0.02115872 0.01608627 0.04376788 0.02817631
#>  [7] 0.04561685 0.02938416 0.03879701 0.04151316

# standard deviation from samples
apply(simplify2array(samples$probs), 1, sd)
#>  [1] 0.04405297 0.04911051 0.02291450 0.02046111 0.04423394 0.03029676
#>  [7] 0.04530734 0.03058053 0.03698335 0.04142802